Spring data merged (#4)
* Spring DATA

Switched to SPRING DATA

* Spring DATA

Switched to SPRING DATA

* Spring DATA

Switched to SPRING DATA
After mentors suggestions

* Spring DATA

Switched to SPRING DATA
After mentors suggestions

* Global Exception Handler and data validation

It is not in main branch but I cant wait too long.

* Global Exception Handler and data validation

It is not in main branch but I cant wait too long.

* Global Exception Handler and data validation

It is not in main branch but I cant wait too long.

* Spring DATA (#3)

* Spring DATA
* Fixed according to Q&A and comments

* Global exception handler and data validation.

Merged previous feature and rebased current branch. Fixed according to Q&A and comments.